+++ apps/backend/src/utils/executors/jsExecutor.ts
@@ -1,30 +1,20 @@
 import { randomUUID } from "node:crypto";
 import fs from 'fs'
-import { spawnSync } from "node:child_process";
 import { TestCase } from "@repo/db";
-import { Outputs } from "src/types/runner.types";
+import spawner from "../spawner.js";
 
-export default function jsExecutor(code: string, testCases: TestCase[]) {
+export default async function jsExecutor(code: string, testCases: TestCase[], timeLimit: number, memoryLimit: number) {
   const uuid = randomUUID()
-  const filename = `${uuid}.py`
+  const filename = `${uuid}.cjs`
   fs.writeFileSync(filename, code, 'utf-8')
 
-  let interpretor = 'node'
-
-  const outputs: Outputs[] = []
-
-  for (const testCase of testCases) {
-    const result = spawnSync(`${interpretor} ${filename}`, {
-      input: testCase.input,
-      encoding: "utf-8"
-    })
-
-    if (result.error) outputs.push({ testCaseId: testCase.id, output: result.stderr, status: "FAIL" })
-    if (result.stdout != testCase.output) outputs.push({ testCaseId: testCase.id, output: result.stdout, status: "FAIL" })
-    else outputs.push({ testCaseId: testCase.id, output: result.stdout, status: "PASS" })
+  try {
+    const response = await spawner(`node`, [`./${filename}`], testCases, timeLimit, memoryLimit)
+    return response
+  } catch (error: any) {
+    console.log(error)
+    return { success: false, message: "Some error occured inside server", error: error.message, errorCode: 500 }
+  } finally {
+    if (fs.existsSync(filename)) fs.unlinkSync(filename)
   }
-
-  fs.unlinkSync(filename)
-
-  return { success: true, data: outputs }
 }
